+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [3.12.0](https://github.com/openshift-hyperfleet/kartograph/compare/v3.11.0...v3.12.0) (2026-02-10)
+
+
+### Features
+
+* **api.iam:** auto tenant root workspace creation ([#210](https://github.com/openshift-hyperfleet/kartograph/issues/210)) ([11db83f](https://github.com/openshift-hyperfleet/kartograph/commit/11db83fa71f74e87fe70d61bb223e1bdabfab96f))
+
+## [3.11.0](https://github.com/openshift-hyperfleet/kartograph/compare/v3.10.0...v3.11.0) (2026-02-10)
+
+
+### Features
+
+* **api.iam:** workspace authz schema alignment ([#208](https://github.com/openshift-hyperfleet/kartograph/issues/208)) ([be1f258](https://github.com/openshift-hyperfleet/kartograph/commit/be1f258108ef526a58a3d9b06e21a7d4a2363096))
+
 ## [3.10.0](https://github.com/openshift-hyperfleet/kartograph/compare/v3.9.0...v3.10.0) (2026-02-09)